feat(dashboard):Remove HomModule
delete extra Module from app
