--- src/Presentation/Presentation.DashBoard/src/app/home/home.routing.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { Routes, RouterModule } from '@angular/router';
-import { AuthGuard } from '../auth/auth.guard';
-
-import { HomeComponent } from './home.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: HomeComponent,
-    canActivate: [AuthGuard],
-  },
-];
-
-@NgModule({
-  imports: [CommonModule, RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class HomeRoutingModule {}
